worker/requirements.txt
celery==3.1.25
requests==2.25.1
PyYAML==3.10
threadpool
